requirements-colab.txt
# Keep Colab's preinstalled GPU-enabled torch.
matplotlib
tokenizers
datasets
numpy
gradio
ebooklib
beautifulsoup4
PyMuPDF
jinja2
transformers
accelerate
sentence-transformers
faiss-cpu
peft
pyyaml
sentencepiece
safetensors
huggingface_hub
bitsandbytes
